# swagger-api-assignment1-r1.yaml
swagger: '2.0'
info:
  description: ''
  version: 1.1.0
  title: Crowdfunding API for SENG365
basePath: /api/v1
...
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /users:
    post:
      tags:
        - users
      summary: Create user
      operationId: createUser
      parameters:
        - in: body
          name: body
          description: User details
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '201':
          description: OK
          schema:
            type: integer
            description: Id of the created user
...
  '/users/{id}':
    get:
      tags:
        - users
      summary: Get user by user id
      description: ''
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/PublicUser'
        '400':
          description: Invalid id supplied
        '404':
          description: User not found
    put:
      tags:
        - users
      summary: Update user
      security:
        - User: []
      description: This can only be done by the logged in user.
      parameters:
        - name: id
          in: path
          description: user id that needs to be updated
          required: true
          type: string
        - in: body
          name: body
          description: Updated user object
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '401':
          description: Unauthorized - not logged in
        '403':
          description: Forbidden - account not owned
        '404':
          description: User not found
    delete:
      tags:
        - users
      summary: Delete user
      security:
        - User: []
      description: This can only be done by the logged in user.
      parameters:
        - name: id
          in: path
          description: The id that needs to be deleted
          required: true
          type: string
      responses:
        '200':
          description: User deleted
        '401':
          description: Unauthorized - not logged in
        '403':
          description: Forbidden - account not owned
        '404':
          description: User not found
definitions:
...
  User:
    type: object
    properties:
      user:
        $ref: '#/definitions/PublicUser'
      password:
        type: string
  PublicUser:
    type: object
    properties:
      id:
        type: integer
      username:
        type: string
      location:
        type: string
      email:
        type: string
...
securityDefinitions:
  User:
    type: apiKey
    name: X-Authorization
    in: header